use int

+++ 005/sol.cpp
@@ -2,13 +2,13 @@
 
 // really slow implementation
 int main() {
-    const uint8_t n = 20;
-    uint8_t count;
+    const int n = 20;
+    int count;
     int i;
 
     for (i = 1; ; i++) {
         count = 0;
-        for (uint8_t j = 1; j <= n; j++) {
+        for (int j = 1; j <= n; j++) {
             if (i % j == 0) count++;
         }
         if (count == n) break;
